Make phrase accept/reject system flag-based
Change the structure of insert and update functions to make phrase
acceptance/rejection flag-based. Essentially, this means consolidating
the InsertPhrase function into one function and making AcceptPhrase and
RejectPhrase update functions that change the flag value of a phrase,
given by the _id of the phrase. These were tested and debugged.